requirements.txt
fastmcp
uvicorn
notion-client
python-dotenv
starlette
pydantic
pinecone
sentence-transformers